Rename butlerQuantumContext.py to _quantumContext.py
* Private naming to match expectations.
* Preparation for renaming ButlerQuantumContext to QuantumContext
