+++ python/lsst/pipe/base/__init__.py
@@ -2,9 +2,9 @@ from . import automatic_connection_constants, connectionTypes, pipelineIR
 from ._dataset_handle import *
 from ._instrument import *
 from ._observation_dimension_packer import *
+from ._quantumContext import *
 from ._status import *
 from ._task_metadata import *
-from .butlerQuantumContext import *
 from .config import *
 from .connections import *
 from .executionButlerBuilder import *
