+++ python/lsst/pipe/base/_quantumContext.py
